backend/requirements.txt
fastapi
uvicorn
ollama
python-dotenv
python-multipart